+++ src/compiler/renderer/ref-renderer.ts
@@ -1,15 +1,56 @@
 import {NodeRenderer} from "./renderer";
-import {Node} from "@unified-latex/unified-latex-types";
+import {Macro, Node} from "@unified-latex/unified-latex-types";
 import {match} from "@unified-latex/unified-latex-util-match";
 import {htmlLike} from '@unified-latex/unified-latex-util-html-like';
 import {classes} from "./classes";
 import {toTagString} from "../../tag";
+import {IRUnit} from "../grouping";
+import {ParserLogger} from "../logging-base";
 
 
 const refCommands = new Set<string>(['ref', 'autoref', 'hyperref']);
 
 
 export class RefRenderer extends NodeRenderer {
+    tagUnitMap: Map<number, IRUnit>;
+    constructor({ tagUnitMap, logger }: {
+        tagUnitMap: Map<number, IRUnit>;
+        logger: ParserLogger;
+    }) {
+        super({logger});
+
+        this.tagUnitMap = tagUnitMap;
+    }
+
+    renderParasitic(node: Macro, target: IRUnit) {
+        console.log("Surely this was called")
+
+        const targetTagString = toTagString(target.tag);
+        const href = `/t/${toTagString(target.parent?.tag ?? 0)}#${targetTagString}`;
+
+        // Essentially, if the ID is already present in the current page, there's no need to navigate to the parent page.
+        // Instead, I should just go to the tag immediately.
+        const fallback = `
+        if (document.getElementById("${targetTagString}")) {
+            location.hash = "${targetTagString}";
+            return false;
+        }
+        `;
+
+        return htmlLike({
+            tag: 'a',
+            attributes: {
+                href, class: classes.ref,
+                onclick: fallback
+            },
+            content: typeof node.refMeta?.text === 'string' ? {
+                type: "string",
+                content: node.refMeta.text
+            } : node.refMeta?.text
+        });
+    }
+
+
     render(node: Node): Node | void {
         if (!(match.anyMacro(node) && refCommands.has(node.content))) return;
         if (!node.refMeta) {
@@ -18,6 +59,13 @@ export class RefRenderer extends NodeRenderer {
         }
 
         if (node.refMeta.targetTag >= 0) {
+            // If the tag refers to a parasitic unit, there will be a special handler.
+            const targetNode = node.refMeta.targetTag ? this.tagUnitMap.get(node.refMeta.targetTag) : undefined;
+
+            if (targetNode?.parasitic && targetNode.parent?.tag !== undefined) {
+                return this.renderParasitic(node, targetNode);
+            }
+
             return htmlLike({
                 tag: 'a',
                 attributes: {