+++ src/compiler/metadata/util.ts
@@ -1,4 +1,4 @@
-import {DisplayMath, Environment, Macro} from "@unified-latex/unified-latex-types";
+import {Argument, DisplayMath, Environment, Macro} from "@unified-latex/unified-latex-types";
 import {match} from "@unified-latex/unified-latex-util-match";
 import {Node} from "@unified-latex/unified-latex-types";
 
@@ -8,3 +8,12 @@ export function isLabelableDisplayMath(node: Node): node is Environment {
     return match.anyEnvironment(node) && node.type === 'mathenv';
 }
 
+export const reservedLabelStart = "auto:";
+
+// The type annotations here lie.
+export function getEnvironmentName(node: Environment) {
+    if (typeof node.env === 'string') return node.env;
+
+    // @ts-ignore
+    return node.env.map((n) => n.type === 'string' ? n.content : '').join('');
+}